<fix>: Build share urls using page url and title I was being stupid, I didn't realize the floats were included in every page, so they could not be hardcoded. There were a few solutions:
Remove the partial, copy and paste the floats code in every page,
changing the url to share. Obviously, not ideal.

Remember to use the partial in every section and either pass the
title, or create a php global variable with it. Problematic, plus
I would have to either hardcode the url or grab it from the page
anyway.

In the end, I decided to grab the url and the title from the request.
I still don't like this solution because it's a bit too clever and
it's not configurable in case the user wants to for some reason.
The blade code and having to create a global helper is ugly, too.