<fix>: Double "http:" in production
I don't understand why, but in production request()->fullUrl() is
returning the url with http:// and in "local" development it is not.

I dove into the code and I see two possible related explanations:

In development I am serving the page using php artisan serve, which
uses the default php -S with a port. I think the port is the key
difference here.

I don't have time to really fix it and the whole thing is a hack.
Best luck in the future.