<fix>: Display mobile header over logos
Because the logos are using a CSS filter to display as grayscale, a
stacking context was being created, so the stacking context created by
the mobile header being sticky was no longer the only one, and the logos
come after the header in the html, so they were displayed on top.

Solved by increasing the z-index of the mobile header, so it displays on
top of the logos, because they are now in the same stacking context.