Remove div container, give its section its own dimensions
The container div avoided repetition, and thus it was faster to work
with, but it was too constraining. I remember wanting the map to be
wider, but couldn't do it. It was not important, so I ingored it.

Now we want the Nuestra experiencia section to be a card, so it has to
be a bit wider, but we can't because of the width of its parent, so the
container has to go.